Boots, Touch, GPS, ***BUGS*** not ready for most users!
This is the first super hacky pass at integrating support for the
LILYGO T-Deck into the ESP32Marauder firmware.

This is not ready for mainstream use. It's just a proof of concept
at this point. I hope to fix enough of the bugs to release it.

Stay tuned! Same channel. Same time.